+++ src/components/attendee/VideoHuddle/internal/HuddleParticipant.tsx
@@ -1,15 +1,5 @@
-import {
-  Participant,
-  VideoSource,
-  VideoTrack,
-} from "components/attendee/VideoComms/types";
+import { Participant, VideoTrack } from "components/attendee/VideoComms/types";
 import { VideoCommsParticipant } from "components/attendee/VideoComms/VideoCommsParticipant";
-
-import { UserId } from "types/id";
-
-import { useProfileById } from "hooks/user/useProfileById";
-
-import { UserAvatar } from "components/atoms/UserAvatar";
 
 import styles from "./HuddleParticipant.module.scss";
 
@@ -24,13 +14,6 @@ export const HuddleParticipant: React.FC<HuddleParticipantProps> = ({
   isLocal = false,
   addButtons,
 }) => {
-  const { profile, isLoading } = useProfileById({
-    userId: participant.sparkleId as UserId,
-  });
-  const hasActiveVideoStream = participant.videoTracks.some(
-    (t) => t.sourceType === VideoSource.Webcam && t.enabled
-  );
-
   return (
     <div key={participant.id} className={styles.huddleParticipant}>
       <VideoCommsParticipant
@@ -38,13 +21,6 @@ export const HuddleParticipant: React.FC<HuddleParticipantProps> = ({
         videoTrackControls={addButtons}
         isLocal={isLocal}
       />
-
-      {!hasActiveVideoStream && !isLoading && (
-        <UserAvatar
-          containerClassName={styles.avatarContainer}
-          user={profile}
-        />
-      )}
     </div>
   );
 };
